feat(server): plugin catch-all via real PluginRegistry (#147 TB18)
Migrates the /api/plugins/:pluginId/:path* catch-all from the Next.js App Router to a Bun-shape handler at packages/host/src/api/plugins/[pluginId]/[[...path]].ts. Roughly 170 lines of duplication removed — the old catch-all had to lazy-activate all 10 plugins in its own isolated Next.js module context because it couldn't share state with server.ts. The new handler reads directly from the real pluginRegistry.

Changes:
- packages/host/src/api/plugins/[pluginId]/[[...path]].ts: single handle() bound to get/post/put/patch/del. Parses /api/plugins/<id>/<subpath> from url.pathname, looks up pluginRegistry.getPluginState(id), runs its matchRoute() over the plugin's registered APIRoutes, builds a per-request ctx (mirroring the old file's buildCtxExtras so handlers still get real activity/audit/hooks/search surfaces), dispatches.
- server.ts: dispatch block before the Next.js fallback; routes /api/plugins/* MINUS the specific install/remove paths already handled above.
- bakin.config.ts: add health plugin. Bug fix surfaced during this migration — the old catch-all activated health in its own context, masking its absence from the main plugin-registry. Now health routes (/api/plugins/health/checks, etc.) flow through the real activation path.

Deleted src/app/api/plugins/ (formerly contained [pluginId]/[[...path]]/route.ts). src/app/api/ is now empty and also removed.

Verification:
- bunx tsc --noEmit clean
- bunx vitest run — 2983 passed, 1 skipped (no regressions from Phase B baseline)
- Runtime smoke: GET /api/plugins/health/checks → 200; GET /api/plugins/workflows/definitions → 200; GET /api/plugins/unknown → 404 with helpful message
